refactor: split cli/interaction.py (2424 lines) into focused submodules (fixes #634)
Split the monolithic interaction.py into 5 focused modules under
naturo/cli/interaction/:

| Module       | Lines | Responsibility                        |
|------------- |-------|---------------------------------------|
| _common.py   |   830 | Shared helpers, decorators, constants |
| _click.py    |   310 | click command                         |
| _type.py     |   370 | type command                          |
| _press.py    |   310 | press and hotkey commands             |
| _mouse.py    |   310 | scroll, drag, move commands           |

Each command module accesses shared helpers via
`import naturo.cli.interaction._common as _common`, enabling a single
consistent mock patch path: `naturo.cli.interaction._common.<name>`.

__init__.py re-exports all public names for backward compatibility.
`from naturo.cli.interaction import click_cmd` and similar imports
continue to work unchanged.

Updated 12 test files (108 mock patch paths) from
`naturo.cli.interaction._<name>` to
`naturo.cli.interaction._common._<name>`.
